[level 2] Title: [PCCP 기출문제] 2번 / 퍼즐 게임 챌린지, Time: 34.07 ms, Memory: 119 MB -BaekjoonHub

+++ 프로그래머스/2/340212. ［PCCP 기출문제］ 2번 ／ 퍼즐 게임 챌린지/［PCCP 기출문제］ 2번 ／ 퍼즐 게임 챌린지.java
@@ -17,7 +17,7 @@ class Solution {
             long sum = times[0];
             for(int idx = 1; idx < size; idx++) {
                 if(mid < diffs[idx]) {
-                    sum += (times[idx] + times[idx - 1]) * (diffs[idx] - mid);
+                    sum += (long) (times[idx] + times[idx - 1]) * (diffs[idx] - mid);
                 }
                 
                 sum += times[idx];
